feat: CoreData의 CRUD중 CR 구현
- 현재 프로젝트 타겟이 하나라 modelName을 주입받는 대신 직접 입력
- CandleData 전용 CoreData타입명인 CandleCoreDataManager로 변경
- CoreData로부터 CandleData를 읽어오는 read(date) 메서드 구현
- CoreData에 CandleData를 추가하는 create() 메서드 구현

+++ CandleData+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension CandleData {
         return NSFetchRequest<CandleData>(entityName: "CandleData")
     }
 
-    @NSManaged public var date: Int64
+    @NSManaged public var date: Double
     @NSManaged public var openPrice: String?
     @NSManaged public var closePrice: String?
     @NSManaged public var highPrice: String?
